Move things into clauses module
