+++ src/program_environment.rs
@@ -1,4 +1,5 @@
 use chalk_ir::ProgramClause;
+
 #[derive(Clone, Debug, PartialEq, Eq)]
 pub struct ProgramEnvironment {
     /// Compiled forms of the above:
